feat(client): implement InsertSnapshot
